Build Atom as a webapp.
The goal of this branch is to create a build script that generates
a version of Atom that can run in a web browser. We have to shim a
bunch of native dependencies as well as calls that are not allowed
from the browser sandbox.

A native dependency, ['buffer-offset-index'](
https://www.npmjs.com/package/buffer-offset-index) was introduced in
`text-buffer@9.4.1`, which was integrated into Atom in
95acc979187b351dc6402c24eddc41d7007e93ec. As such, we are basing this
work on the parent commit (f7d3f0210bf6ff1b4193d8a8b8a54c199b561bc2)
so that we have a reasonably up-to-date version of Atom as
a starting point.