Add a browserify transform to replace require.resolve() calls.
We have been having an issue getting the `find-and-replace` Atom package
to load cleanly because of the following code in the `atom-space-pen-views`
Node module:

```
atom.themes.requireStylesheet(require.resolve('../stylesheets/select-list.less'))
```

The problem is that Browserify does not rewrite the `require.resolve()`
into something that works at runtime in the browser.

This commit introduces a transform that converts that code into the following:

```
atom.themes.requireStylesheet(
  'absolute-path-to/node_modules/atom-space-pen-views/stylesheets/select-list.less'
)
```

and ensures that the corresponding `select-list.less` file gets a
proper entry in the in-memory filesystem on the client.

This change does most of the work to unblock support for the
`find-and-replace` package (which is now enabled in addition to `tabs`,
even though it doesn't work as of this commit). Now the problem is that the
`LessCompileCache` on the client is not initialized properly, which will
be fixed in the subsequent commit.