Include static/ directory with .less files and load via LessCompileCache.
This takes the `static/` folder from the source checkout of Atom and
makes it available under `/Users/zuck/resourcePath/static` on the
client and tells the client that `/Users/zuck/resourcePath` is the
`resourcePath` to use as part of the parameters that are passed to
the window via the URL hash.

This is important because the `ThemeManager` receives this path and
passes it to the `LessCompileCache` cache that it creates for itself.

Now that `LessCompileCache` is loaded correctly, the client loads
without error with the `find-and-replace` package activated.

However, for some reason that I have not figured out yet, `cmd-f` does not bring
up the find-and-replace GUI (the browser's built-in find-and-replace comes up
instead). Presumably the keymap is not being loaded correctly (or Chrome is
trapping that event in an unfortunate way).

Nevertheless, running the following from the console in Chrome DevTools does
bring up the GUI for `find-and-replace`:

```
atom.commands.dispatch(atom.views.getView(atom.workspace), 'find-and-replace:show')
```

This change should also help us remove the hardcoded `styles.css` file,
but I'll attempt that in a follow-up commit. (Admittedly, the CSS looks
a little funny now, but I suspect that's because we're also loading a
combination of `styles.css` and the contents of the `static` directory now.)