Setting process.platform correctly fixes some key issues.
I dug into why `cmd-f` was not loading `find-and-replace`. The issue
was that `process.platform` was set to `'browser'` by default by
Browserify, which is a big deal for Atom.

On OS X, Atom adds the `platform-darwin` CSS class to the `<body>` tag,
which is critical because keybindings target CSS selectors in Atom.
When `process.platform` was set to `'browser'`, the wrong CSS class was
set on `<body>`, so the `cmd-f` shortcut did not trigger, which
prevented `find-and-replace` from activating.

I added code to check `window.navigator.platform` and set
`process.platform` appropriately. I also amended the list of resource
directories to include `keymaps`, `menus`, and `resources`.
Now that we load `keymaps` in the standard way, I was able to
delete a bunch of the `STANDALONE_KEYMAP` stuff we had before. Also,
whatever special stuff we were doing to make cut/copy/paste work no
longer seems necessary (at least on Chrome 53.0.2785.143, though I
know James pointed out some keybinding fixes that we may have to
pick up on Atom master for things to continue to work).

Getting rid of a bunch of stuff we were special-casing in this
commit makes me very happy :)