Enable more Atom packages by default.
Added the following packages:
* `command-palette`
* `go-to-line`
* `markdown-preview`
* `status-bar`

In general, everything seems to work, except when
the markdown preview pane opens for the untitled file, it says:

"Previewing Markdown Failed."

This could be related to our general issue with fs.createReadStream()
not working. Will still enable for now until we dig into the BrowserFS
issues and fix things such that we can read and write files.

There's also a complaint about `octocat-spinner-128.gif` not loading
in the console.