Have atom.js export a load() function rather than call it.
Now, after `atom.js` loads, a `loadAtom()` function will be available in
the global scope. It takes a callback that receives one argument,
which is an object with properties on it. To start, the only available
properties are the `atom` global and `fs` (which is really `fs-plus`,
but it is still backed by BrowserFS.)

The `testpage.html` demo page has been updated to demonstrate how this
works. In particular, note how it calls `loadAtom()` and then calls
`atom.addPath()` itself. Other users may want to start writing files
to BrowserFS via `fs` before calling `atom.addPath()`.